+++ app/src/main/java/ch/epfl/sdp/cook4me/ui/eventform/Event.kt
@@ -54,12 +54,19 @@ data class Event(
     constructor(map: Map<String, Any>) : this(
         name = map["name"] as? String ?: "",
         description = map["description"] as? String ?: "",
-        dateTime = (map["dateTime"] as? com.google.firebase.Timestamp)
+        dateTime = (map["dateTime"] as? Map<String, Any>)
+            ?.let { it["time"] as? com.google.firebase.Timestamp }
             ?.toDate()
             ?.let { calendarFromTime(it) }
             ?: Calendar.getInstance(),
         location = map["location"] as? String ?: "",
-        maxParticipants = map["maxParticipants"] as? Int ?: 0,
+        /*
+        * map["maxParticipants"]: get the value of the key "maxParticipants" in the map
+        * map["maxParticipants"] as? Long: cast the value to a Long, if it is not possible, return null
+        * (map["maxParticipants"] as? Long)?.toInt(): if the value is not null, cast it to an Int
+        * (map["maxParticipants"] as? Long)?.toInt() ?: 0 : provide a default value if the value is null (0)
+        * */
+        maxParticipants = (map["maxParticipants"] as? Long)?.toInt() ?: 0,
         participants = map["participants"] as? List<String> ?: listOf(),
         creator = map["creator"] as? String ?: "",
         id = map["id"] as? String ?: "",
